+++ src/continuity/trainer/trainer.py
@@ -80,18 +80,18 @@ class Trainer:
             dataset: Data set.
             tol: Tolerance for stopping criterion. Ignored if criterion is not None.
             epochs: Maximum number of epochs.
-            callbacks: List of callbacks. Defaults to [PrintTrainingLoss] if verbose.
+            callbacks: List of callbacks. Defaults to PrintTrainingLoss, if verbose.
             criterion: Stopping criterion. Defaults to TrainingLossCriteria(tol).
             batch_size: Batch size.
             shuffle: Shuffle data set.
             val_dataset: Validation data set.
         """
-        # Default callback
+        # Default callbacks
         if callbacks is None:
+            callbacks = []
+
             if self.verbose:
-                callbacks = [PrintTrainingLoss()]
-            else:
-                callbacks = []
+                callbacks.append(PrintTrainingLoss())
 
         # Default criterion
         if criterion is None:
@@ -99,7 +99,7 @@ class Trainer:
 
         # Print number of model parameters
         if self.verbose:
-            num_params = sum(p.numel() for p in self.operator.parameters())
+            num_params = self.operator.num_params()
             print(f"Model parameters: {num_params}")
 
         # Move operator to device
@@ -184,3 +184,6 @@ class Trainer:
 
         # Move operator back to CPU
         self.operator.to("cpu")
+
+        # Return statistics
+        return {"loss/train": loss_train, "loss/val": loss_val, "epoch": epoch + 1}