+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - Change `dataptation` dependency from "==3.1" to ">=3.1,<4.0".
 - Change `optuna` dependency from "3.5.0" to ">=3.5.0,<4.0.0".
 - Add `FourierLayer` and `FourierNeuralOperator` with example.
+- Add `benchmarks` infrastructure.
 
 ## 0.0.0 (2024-02-22)